Empty names after reload
Fixed a bug on plugin reload where temporary permissions and temporary groups would have their name become an empty string (Thanks to Cerridwen for finding it).
Removed the revoke of temporary permissions/groups on player disconnect and re-grant on connect. Now this only happens during plugin unload or load, if the corresponding config is true.