0.1.7
Added the ability to assign temporary permissions and temporary groups to players who have not yet connected to the server. For commands, this only works if a SteamID is specified (so it can be validated);
Moved registration of commands to the end of the plugin initialization, due to NRE errors caused by commands being called before full initialization;
Other minor changes and code improvements.